+++ source/redux/server/store.js
@@ -2,7 +2,7 @@ import { reduxReactRouter } from 'redux-router/server'
 import createHistory from 'history/lib/createMemoryHistory'
 import create_store from '../store'
 
-export default function create_store_on_server(get_reducer, settings)
+export default function create_store_on_server(reducer, settings)
 {
-	return create_store(reduxReactRouter, createHistory, get_reducer, settings)
+	return create_store(reduxReactRouter, createHistory, reducer, settings)
 }