+++ source/render on server.js
@@ -5,7 +5,7 @@ import ReactDOMServer from 'react-dom/server'
 // to the resulting Html markup code
 // (returns a string containing the final html markup)
 //
-export default function render_on_server({ render_webpage_as_react_element, page_element })
+export default function render_on_server({ render_webpage, page_element })
 {
-	return '<!doctype html>\n' + ReactDOMServer.renderToString(render_webpage_as_react_element(page_element))
+	return '<!doctype html>\n' + render_webpage(page_element)
 }